addes places to files

+++ middlewares/validations/post/postEditValidation.js
@@ -1,18 +1,14 @@
-const deleteFile = require('../../../api/utils/deleteFile')
-const { serverError } = require('../../../constants/index')
+const deleteFile = require("../../../api/utils/deleteFile")
+const { serverError } = require("../../../constants/index")
 
 const postEditValidation = async (req, res, next) => {
   const { data } = req.body
   const maxDataLength = 1000
 
-  function handleBadRequest(errorCode, message){
+  function handleBadRequest(errorCode, message) {
     /* Delete previous files */
-    if (req.files) {
-      for(let i in req.files){
-        deleteFile(req.files[i].key)
-      }
-    }
-  
+    if (req.files) for (let i in req.files) deleteFile(req.files[i].key)
+
     return res.status(errorCode).json({ message })
   }
 
@@ -21,7 +17,7 @@ const postEditValidation = async (req, res, next) => {
       return handleBadRequest(413, `Text is too long`)
 
     return next()
-  }catch(error){
+  } catch (error) {
     return handleBadRequest(500, serverError(error.message).message)
   }
 }
